Add a webpack definition so that the module "component-uid" doesn't require "crypto-browserify"
component-uid does a check to see if "window" is defined. If it's not, then it requires "crypto", which is a
NodeJS built-in (but if has a *big* polyfill for the browser). By resolving that conditional
at compilation time, the module is not required.
This saves 300 KB of the JS bundle.