Enable babel's "loose mode"
Loose mode (for the transforms that support it) make the transform not conform 100% to
the specification, so it could break in really obscure ways. The upside is that loose-transformed
code is smaller and faster because it doesn't need to take into account those edge cases.
I've looked through all the transforms that have loose mode and we are fine. These are them:
https://babeljs.io/docs/plugins/transform-es2015-classes/
https://babeljs.io/docs/plugins/transform-es2015-computed-properties/
https://babeljs.io/docs/plugins/transform-es2015-for-of/
https://babeljs.io/docs/plugins/transform-es2015-modules-commonjs/
https://babeljs.io/docs/plugins/transform-es2015-spread/
https://babeljs.io/docs/plugins/transform-es2015-template-literals/

This saves 9KB of the JS bundle, which is not a lot, but it will also improve performance slightly.