Show TOS-only banner if connect button not clicked
Before, we were recording whether to show the TOS-only
banner when the before banner was displayed.

Now, we record whether or not display it after the button
in the before banner is clicked.

This prevents the case that the user sees the before banner,
does not click it, goes through the Jetpack connection process
(not the WooCommerce Services connect process), and then sees
the after connection banner and has their TOS accepted.

Instead, we want this user to see the TOS-only banner to
explicitly accept the TOS using the button on the TOS-only banner.