Fix "undefined JETPACK__VERSION" warning
Fixes #1241

Even when Jetpack isn't installed, the admin_enqueue_scripts function is run, so we can't assume that the `JETPACK__VERSION` constant is available. Since it's only used to put a cache-bustin querystring on the `noticons.css` URL, it's safe to replace with a dummy value if Jetpack is not available.

To test:
* Make sure `WP_DEBUG` is `TRUE`.
* Disable or uninstall Jetpack.
* Check that there's no `Undefined constant JETPACK__VERSION` PHP warning after visiting any WP-Admin page.