Allow changing label settings (except payment) to non-Master users
Changed the `label settings` endpoint logic so any admin user is allowed to update the settings.

If the user is not the `Jetpack master user`, then the payment setting won't be updated, but the rest of the settings will.

It also adds a flag to the `GET` endpoint (`can_edit_settings`) which will always be `true`. This is done so the Calypso UI can differentiate between versions of the endpoints:
* For previous versions, `can_edit_settings` will be `false`-y (`undefined`), so only the `Master` user will be able to update the settings.
* For versions that include the `can_edit_settings` flag in the `GET` response, it's safe to present the user with the `Settings` form, and only hide the "Payment" part of it if he's not the `Master` user.