Fixes the URL to print a label when the "Permalink" setting is off
Fixes #1260

The bug was introduced in `1.9.0`, when we hybridized the extension. The plugin code is using the `api.js` module from `wp-calypso`, and that handles URL structures differently.

This is the minimal change, I have prepared a more comprehenxsive fix that cleans up some code, but I figured I should leave that for the release after we unfreeze the code.

To test:
* Go to `Settings -> Permalinks` and select the option `Plain`.
* Try to purchase or reprint a label. It should work.