Updated NodeJS to 8.9.3
Node v8 is now the LTS version, and Calypso has started using it a few weeks ago.

This also changes `npm` to version 5, which is much faster and reliable. `npm v5` comes bundled with `nodejs v8`, but if for some reason it isn't working for you, just type `npm i -g npm@5` to install it.

This PR completely regenerates the `shrinkwrap` file, so it's not risk-free. I didn't find anything obviously broken though.