Respond with real HTTP errors when the shipping settings can't be validated (#1428)
* Respond with real HTTP errors when the shipping settings can't be validated

There's more context in https://github.com/Automattic/wp-calypso/pull/25101

We had an error in which service schema validation would fail in the server if the adjustment fields were touched in any way. The error has been fixed (something something `coerceValues`), but I dug deeper on "Why doesn't this fail in Calypso"?

In Calypso, a validation failure in the plugin side will happily show a "success" notice, but the shipping method won't be updated at all. That's because the Calypso HTTP layer logic checks for non-200 HTTP responses to trigger the error notice, and the plugin responds with a `HTTP 200` response with `{ success: false }` in the body.

This PR changes the plugin logic so it responds with a real HTTP error. All our other REST endpoints already return a `WP_Error` on failure, this one was the only one that used the `{ success: false }` trick.

To test:
- With this branch of the plugin installed on your site, go to Calypso (can be `wordpress.com` or a local install).
- Go to the Shipping settings, and add or edit a USPS shipping method.
- Before saving, stop your local Hydra server (or make the `WOOCOMMERCE_CONNECT_SERVER_URL` constant to point to a non-existant server).
- Save the changes in Calypso.
- Check that you get an error notice. Before, it would just fail silently and display a Success notice.

* Fixed warning when WP_Error->get_error_data returns NULL