Fixes the "Save changes" button on USPS/Canada settings staying in "busy" state after failing validation (#1440)
* Fixes the "Save changes" button on USPS/Canada settings staying in "busy" state after failing validation

If there are invalid fields when you open the USPS/Canada settings form, but you haven't interacted with those fields, then they won't show any error until you click "Save changes". Clicking that button will trigger a full validation and show an error notice if it fails (the logic of that is in `wp-calypso`, including the error notice). Since all that logic is in `wp-calypso`, the `<ViewWrapper>` component's state wasn't being updated, so it would keep showing the `busy` animation forever.

To reproduce:
* Get to the service settings page having some fields be incorrect. The easiest way is to change the Store base address to have an invalid ZIP code, and then add a new shipping method (it will take that bogus ZIP code as default).
* Interact with any other field (such as the title, or services list).
* Click the `Save changes` button.
* In `master`, you'll see the error notice but the button will keep being "busy". With this change, it will become disabled but it will be clickable again once you fix the errors.

* Added comments to clarify the logic