Add error on checkout only if user can edit address on cart
When a WC error notice is added on checkout, the customer cannot see
the checkout page. Instead the customer will see the "there are some
issues with the items in your cart" message - i.e. the checkout/cart-errors.php page.

When the `woocommerce_enable_shipping_calc` is set to "no", the user cannot edit
the address, so in this case we want to make sure that we do not add
any error notices so that the user can navigate to checkout.