Ignore county when country is UK for automated tax
When country is UK, the shipping address allow county to be a text field.
This text field does not respect a "state code" format (ie. 2 letters).
This creates a problem with WooCommerce core's tax code. While the core
team is working on the issue, we are going to patch this by disabling
county in the tax field for automated tax. County is optional and
this change will not break tax rate.