Catch exception if DHL available note cannot be created.
It is possible for WC Admin to be available but disabled if the
`woocommerce_admin_disabled` filter is used. If this happens,
attempts to create a note will fail due to the "admin-note" data store
being unavailable.

This commit adds a check if an exception is thrown on note creation
and removal attempts.