Check that $order->get_date_created() is not null before calling the getTimestamp() method.
Previously the code assumed that `$order->get_date_created()` would always return an object with the `getTimestamp()` method. If an order is still an auto-draft, however, the `date_created` order meta does not yet exist and `get_date_created()` will return `null`.

This commit checks to see whether or not`$order->get_date_created()` returns a truthy value before attempting to call `getTimestamp()`. If not, it assigns it as `0` (int) since `getTimestamp()` returns an int - that way we're not changing the type of `created_at`.