Adds a test to make sure that get_order_for_api returns expected values when certain order meta values are null.
In `WC_Connect_Order_Presenter->get_order_for_api()`, the `date_created`, `date_modified`, and `date_completed` order meta values are expected to return a `WC_DateTime` object so that `getTimestamp()` can be called on them. We're testing that if any of those metas are `null` for any reason that values still get assigned for the order data returned.